Add a bare-bones VirtualMachine class to the WASM library
