+++ wasm/lib/Cargo.toml
@@ -10,8 +10,8 @@ repository = "https://github.com/RustPython/RustPython/tree/master/wasm/lib"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-rustpython_parser = {path = "../../parser"}
-rustpython_vm = {path = "../../vm"}
+rustpython_parser = { path = "../../parser" }
+rustpython_vm = { path = "../../vm" }
 cfg-if = "0.1.2"
 wasm-bindgen = "0.2"
 js-sys = "0.3"
