+++ wasm/lib/src/lib.rs
@@ -1,3 +1,5 @@
+mod convert;
+mod vm_class;
 mod wasm_builtins;
 
 extern crate js_sys;
@@ -5,111 +7,17 @@ extern crate rustpython_vm;
 extern crate wasm_bindgen;
 extern crate web_sys;
 
-use js_sys::{Array, Object, Reflect, TypeError};
+use js_sys::{Object, Reflect, TypeError};
 use rustpython_vm::compile;
-use rustpython_vm::pyobject::{self, PyFuncArgs, PyObjectRef, PyResult};
+use rustpython_vm::pyobject::{PyFuncArgs, PyObjectRef, PyResult};
 use rustpython_vm::VirtualMachine;
-use wasm_bindgen::{prelude::*, JsCast};
+use wasm_bindgen::prelude::*;
 
-// Hack to comment out wasm-bindgen's typescript definitons
+pub use vm_class::*;
+
+// Hack to comment out wasm-bindgen's generated typescript definitons
 #[wasm_bindgen(typescript_custom_section)]
 const TS_CMT_START: &'static str = "/*";
-
-fn py_str_err(vm: &mut VirtualMachine, py_err: &PyObjectRef) -> String {
-    vm.to_pystr(&py_err)
-        .unwrap_or_else(|_| "Error, and error getting error message".into())
-}
-
-fn py_to_js(vm: &mut VirtualMachine, py_obj: PyObjectRef) -> JsValue {
-    let dumps = rustpython_vm::import::import(
-        vm,
-        std::path::PathBuf::default(),
-        "json",
-        &Some("dumps".into()),
-    )
-    .expect("Couldn't get json.dumps function");
-    match vm.invoke(dumps, pyobject::PyFuncArgs::new(vec![py_obj], vec![])) {
-        Ok(value) => {
-            let json = vm.to_pystr(&value).unwrap();
-            js_sys::JSON::parse(&json).unwrap_or(JsValue::UNDEFINED)
-        }
-        Err(_) => JsValue::UNDEFINED,
-    }
-}
-
-fn js_to_py(vm: &mut VirtualMachine, js_val: JsValue) -> PyObjectRef {
-    if js_val.is_object() {
-        if Array::is_array(&js_val) {
-            let js_arr: Array = js_val.into();
-            let elems = js_arr
-                .values()
-                .into_iter()
-                .map(|val| js_to_py(vm, val.expect("Iteration over array failed")))
-                .collect();
-            vm.ctx.new_list(elems)
-        } else {
-            let dict = vm.new_dict();
-            for pair in Object::entries(&Object::from(js_val)).values() {
-                let pair = pair.expect("Iteration over object failed");
-                let key = Reflect::get(&pair, &"0".into()).unwrap();
-                let val = Reflect::get(&pair, &"1".into()).unwrap();
-                let py_val = js_to_py(vm, val);
-                vm.ctx
-                    .set_item(&dict, &String::from(js_sys::JsString::from(key)), py_val);
-            }
-            dict
-        }
-    } else if js_val.is_function() {
-        let func = js_sys::Function::from(js_val);
-        vm.ctx.new_rustfunc(
-            move |vm: &mut VirtualMachine, args: PyFuncArgs| -> PyResult {
-                let func = func.clone();
-                let this = Object::new();
-                for (k, v) in args.kwargs {
-                    Reflect::set(&this, &k.into(), &py_to_js(vm, v))
-                        .expect("Couldn't set this property");
-                }
-                let js_args = Array::new();
-                for v in args.args {
-                    js_args.push(&py_to_js(vm, v));
-                }
-                func.apply(&this, &js_args)
-                    .map(|val| js_to_py(vm, val))
-                    .map_err(|err| js_to_py(vm, err))
-            },
-        )
-    } else if let Some(err) = js_val.dyn_ref::<js_sys::Error>() {
-        let exc_type = match String::from(err.name()).as_str() {
-            "TypeError" => &vm.ctx.exceptions.type_error,
-            "ReferenceError" => &vm.ctx.exceptions.name_error,
-            "SyntaxError" => &vm.ctx.exceptions.syntax_error,
-            _ => &vm.ctx.exceptions.exception_type,
-        }
-        .clone();
-        vm.new_exception(exc_type, err.message().into())
-    } else if js_val.is_undefined() {
-        // Because `JSON.stringify(undefined)` returns undefined
-        vm.get_none()
-    } else {
-        let loads = rustpython_vm::import::import(
-            vm,
-            std::path::PathBuf::default(),
-            "json",
-            &Some("loads".into()),
-        )
-        .expect("Couldn't get json.loads function");
-
-        let json = match js_sys::JSON::stringify(&js_val) {
-            Ok(json) => String::from(json),
-            Err(_) => return vm.get_none(),
-        };
-        let py_json = vm.new_str(json);
-
-        vm.invoke(loads, pyobject::PyFuncArgs::new(vec![py_json], vec![]))
-            // can safely unwrap because we know it's valid JSON
-            .unwrap()
-    }
-}
 
 fn base_scope(vm: &mut VirtualMachine) -> PyObjectRef {
     let builtins = vm.get_builtin_scope();
@@ -182,7 +90,7 @@ pub fn eval_py(source: &str, options: Option<Object>) -> Result<JsValue, JsValue
                             &JsValue::UNDEFINED,
                             &wasm_builtins::format_print_args(vm, args)?.into(),
                         )
-                        .map_err(|err| js_to_py(vm, err))?;
+                        .map_err(|err| convert::js_to_py(vm, err))?;
                         Ok(vm.get_none())
                     },
                 )
@@ -208,7 +116,7 @@ pub fn eval_py(source: &str, options: Option<Object>) -> Result<JsValue, JsValue
             let pair = pair?;
             let key = Reflect::get(&pair, &"0".into()).unwrap();
             let val = Reflect::get(&pair, &"1".into()).unwrap();
-            let py_val = js_to_py(&mut vm, val);
+            let py_val = convert::js_to_py(&mut vm, val);
             vm.ctx.set_item(
                 &injections,
                 &String::from(js_sys::JsString::from(key)),
@@ -220,8 +128,8 @@ pub fn eval_py(source: &str, options: Option<Object>) -> Result<JsValue, JsValue
     vm.ctx.set_item(&mut vars, "js_vars", injections);
 
     eval(&mut vm, source, vars)
-        .map(|value| py_to_js(&mut vm, value))
-        .map_err(|err| py_str_err(&mut vm, &err).into())
+        .map(|value| convert::py_to_js(&mut vm, value))
+        .map_err(|err| convert::py_str_err(&mut vm, &err).into())
 }
 
 #[wasm_bindgen(typescript_custom_section)]
